[refactor] eslint 설정 변경 및 코드 리팩토링
1. import/no-cycle
- index.js 삭제

기타 코드 스타일 수정
