+++ src/components/atoms/SearchResultListItem/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { StarCount } from '@atoms';
+import StarCount from '@atoms/StarCount';
 import './style.scss';
 
-export default function SearchListItem({ name, starCount }) {
+export default function SearchResultListItem({ name, starCount }) {
   return (
     <div className="search_result_item">
       <div className="item_name">{name}</div>
