--- src/components/atoms/index.js
@@ -1,3 +1,0 @@
-export { default as Button } from './Button';
-export { default as SearchResultListItem } from './SearchResultListItem';
-export { default as StarCount } from './StarCount';
