+++ src/components/molecules/SearchResultDescription/index.jsx
@@ -10,8 +10,14 @@ export default function SearchResultDescription({
     <div className="search_result_description">
       <div className="descript_name">{name}</div>
       <div className="descript_count_wrapper">
-        <span className="repo_count">{repoCount} Repositories</span>
-        <span className="star_count">{starCount} stars</span>
+        <span className="repo_count">
+          {repoCount}
+          Repositories
+        </span>
+        <span className="star_count">
+          {starCount}
+          stars
+        </span>
       </div>
     </div>
   );
