--- src/components/molecules/index.js
@@ -1,3 +1,0 @@
-export { default as SearchBar } from './SearchBar';
-export { default as SearchResultDescription } from './SearchResultDescription';
-export { default as SearchResultList } from './SearchResultList';
