--- src/components/organisms/index.js
@@ -1,2 +1,0 @@
-export { default as GitStarRanking } from './GitStarRanking';
-export { default as SearchResult } from './SearchResult';
